requirements.txt
streamlit
python-dotenv
praw
langchain
langchain-core
langchain-groq
langchain-ollama
